+++ src/Agent/NewRelic/Agent/Core/Transactions/Transaction.cs
@@ -34,6 +34,7 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using NewRelic.Agent.Core.OpenTelemetryBridge;
+using NewRelic.Agent.Extensions.Api.Experimental;
 
 namespace NewRelic.Agent.Core.Transactions
 {
