+++ tests/Agent/UnitTests/Core.UnitTest/OpenTelemetryBridge/NewRelicActivitySourceProxyTests.cs
@@ -8,6 +8,7 @@ using System.Reflection;
 using System.Threading;
 using NewRelic.Agent.Api;
 using NewRelic.Agent.Core.Segments;
+using NewRelic.Agent.Extensions.Api.Experimental;
 using NUnit.Framework;
 using Telerik.JustMock;
 using Telerik.JustMock.Helpers;
